backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Giulio Indennitate <giulio.indennitate@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "5.1.1"
djangorestframework = "3.15.2"
django-cors-headers = "4.4.0"
drf-standardized-errors = "0.14.1"
drf-spectacular = "0.27.2"
django-filter = "24.3"
psycopg2-binary = "2.9.9"
structlog = "24.4.0"
django-structlog = "8.1.0"
rich = "13.8.1"
pydantic-settings = "2.6.1"
django-health-check = "3.18.3"


[tool.poetry.group.dev.dependencies]
django-stubs = "5.1.0"
djangorestframework-stubs = "3.15.1"
mypy = "1.11.2"
pytest = "8.3.3"
pytest-mock = "3.14.0"
pytest-cov = "5.0.0"
pytest-django = "4.9.0"
pytest-sugar = "1.0.0"
ruff = "0.7.3"
model-bakery = "1.20.0"


[tool.poetry.group.ci.dependencies]
tox = "4.23.2"
tox-docker = "5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
include = [
    "{backend,prisma_api,core,edge,tests}/**/*.{py,pyi}",
    "pyproject.toml",
]
lint.exclude = ["**/*/migrations/**/*.*"]
lint.select = [
    "A",   # flake8-builtins
    "B",   # flake8-bugbear
    "BLE", # flake8-blind-except
    "C4",  # flake8-comprehensions
    "C90", # mccabe
    "D",   # pydocstyle
    "DJ",  # flake8-django
    "DTZ", # flake8-datetimez
    "E",   # pycodestyle errors
    "ERA", # eradicate
    "EXE", # flake8-executable
    "F",   # pyflakes
    "G",   # flake8-logging-format
    "I",   # isort
    "ICN", # flake8-import-conventions
    "ISC", # flake8-implicit-str-concat
    "N",   # pep8-naming
    "PIE", # flake8-pie
    "PLC", # pylint - convention
    "PLE", # pylint - error
    "PLW", # pylint - warning
    "PTH", # flake8-use-pathlib
    "Q",   # flake8-quotes
    "RET", # flake8-return
    "RUF", # Ruff-specific rules
    "S",   # flake8-bandit
    "SIM", # flake8-simplify
    "T10", # flake8-debugger
    "T20", # flake8-print
    "TID", # flake8-tidy-imports
    "UP",  # pyupgrade
    "W",   # pycodestyle - warning
    "YTT", # flake8-2020
]
line-length = 120
lint.ignore = [
    "A003",    # flake8-builtins - class attribute {name} is shadowing a python builtin
    "B010",    # flake8-bugbear - do not call setattr with a constant attribute value
    "D100",    # pydocstyle - missing docstring in public module
    "D104",    # pydocstyle - missing docstring in public package
    "D105",    # pydocstyle - missing docstring in magic method
    "D106",    # pydocstyle - missing docstring in public nested class
    "E501",    # pycodestyle line too long, handled by ruff format
    "PLW2901", # pylint - for loop variable overwritten by assignment target
    "RUF012",  # Ruff-specific rule - annotated with classvar
    "ISC001",  # Ruff formatter incompatible
]
src = ["backend", "core", "tests"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["backend", "core", "tests"]

[tool.ruff.lint.per-file-ignores]
"**/*tests/**/*.*" = [
    "A",
    "ARG",
    "B",
    "BLE",
    "C901",
    "D",
    "DTZ",
    "EM",
    "FBT",
    "G",
    "N",
    "PGH",
    "PIE",
    "PLR",
    "PLW",
    "PTH",
    "RSE",
    "S",
    "S101",
    "SIM",
    "TCH",
    "TRY",
    "E721",
]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 88

[tool.mypy]
files = ["core"]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = "^(migrations|.*migrations/).*"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main", "pydantic.mypy"]

[tool.django-stubs]
django_settings_module = "backend.settings"

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "backend.settings"
ignore_errors = true

[[tool.mypy.overrides]]
module = "backend.asgi"
ignore_errors = true

[[tool.mypy.overrides]]
module = "backend.wsgi"
ignore_errors = true

[tool.pytest.ini_options]
addopts = [
    "-ra",
    "--cov=backend",
    "--cov=core",
    "--cov-report=term-missing",
    "--create-db",
    "--reuse-db",
]
DJANGO_SETTINGS_MODULE = "backend.settings"
testpaths = ["tests"]
markers = ["unit: mark a unit test.", "integration: mark an integration test."]
filterwarnings = [
    "ignore::DeprecationWarning:importlib._bootstrap",
    "ignore::DeprecationWarning:pkg_resources",
]

[tool.coverage.report]
exclude_lines = [
    'if TYPE_CHECKING:',
    'pragma: no cover',
    "if __name__ == .__main__.:",
    'def __repr__',
    'if self\.debug:',
    'if settings\.DEBUG',
    'raise AssertionError',
    'raise NotImplementedError',
    'if 0:',
    'class .*\bProtocol\):',
    '@(abc\.)?abstractmethod',
]
omit = ["*/tests/*"]
show_missing = true

[tool.coverage.run]
omit = [
    "**/*/__init__.py",
    "**/*/asgi.py",
    "**/*/wsgi.py",
    "**/*/settings.py",
    "**/*/migrations/**/*.py",
    "**/*/manage.py",
]
